refactor: updated module structure to allow supporting multiple ietf draft standards
