+++ src/spike/defaults/versions/latest/index.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = require('../v00')
