+++ src/spike/defaults/versions/v00/constants.js
@@ -1,0 +1,10 @@
+'use strict'
+
+module.exports = {
+  HEADERS_FIELD_PREFIX: 'headers=',
+  SIGNATURE_FIELD_PREFIX: 'signature=',
+  KEY_ID_FIELD_PREFIX: 'keyId=',
+  ALGORITHM_FIELD_PREFIX: 'algorithm=',
+  CREATED_FIELD_PREFIX: 'created=',
+  EXPIRES_FIELD_PREFIX: 'expires='
+}
