+++ src/spike/defaults/versions/v00/constructSignature.js
@@ -1,5 +1,7 @@
+'use strict'
+
 const { createHmac } = require('crypto')
-const { BASE_64 } = require('./constants')
+const { BASE_64 } = require('../../../constants')
 
 module.exports = function constructSignature (secret, algorithm, input) {
   return createHmac(algorithm, secret).update(input).digest(BASE_64).toString()
