+++ src/spike/defaults/versions/v00/index.js
@@ -1,0 +1,17 @@
+'use strict'
+
+const constructSignature = require('./constructSignature')
+const constructSignatureString = require('./constructSignatureString')
+const outputHandler = require('./outputHandler')
+const transformers = require('./transformers')
+const extractors = require('./extractors')
+const signatureFields = require('./signatureFields')
+
+module.exports = {
+  constructSignatureString,
+  constructSignature,
+  outputHandler,
+  transformers,
+  extractors,
+  signatureFields
+}
