+++ src/spike/defaults/versions/v00/signatureFields.js
@@ -1,0 +1,8 @@
+'use strict'
+
+const signatureFields = [
+  '(request-target)',
+  'Date'
+]
+
+module.exports = signatureFields
